+++ packages/auth/package.json
@@ -27,6 +27,8 @@
     "main": "./dist/index.js",
     "types": "./dist/index.d.ts",
     "peerDependencies": {
+        "expo-auth-session": ">=5.0",
+        "expo-crypto": ">=12.4",
         "react": ">=18.0",
         "react-dom": ">=18.0",
         "react-native": ">=0.70",
@@ -34,8 +36,6 @@
     },
     "dependencies": {
         "@equinor/mad-components": "*",
-        "expo-auth-session": "~5.0.2",
-        "expo-crypto": "~12.4.1",
         "jwt-decode": "^4.0.0"
     },
     "devDependencies": {
